moi-main/requirements.txt
ultralytics
opencv-python
mediapipe
pandas
tqdm
pyyaml
torch
torchvision
torchaudio
scikit-learn